### 2018-06-15 #779 @whedon
[PRE REVIEW]: BAMnostic: an OS-agnostic toolkit for genomic sequence analysis
**Submitting author:** @betteridiot (<a href="http://orcid.org/0000-0002-0243-4609">Marcus</a>)
**Repository:** <a href="https://github.com/betteridiot/bamnostic" target ="_blank">https://github.com/betteridiot/bamnostic</a>
**Version:** 0.8.4
