### 2018-06-15 #779 @betteridiot
Unofficially: I love that the JOSS bot is called whedon.
