### 2018-06-15 33c9485
reduced object creation in cigar_align
