### 2018-06-16 #779 @betteridiot
As instructed, using the provided list of potential reviewers, I think the following would be well suited and qualified if they were selected to review this submission (in no particular order):
@ctb
@afrubin
@conradstack
@jean997
@wdecoster
@Adrianzo
@luizirber
