### 2018-06-17 #779 @betteridiot
For the sake of complete transparency, a portion of my BgzfReader is a heavily modified version of the BGZF reader that @peterjc wrote. There is even attribution within that module to it. Furthermore, early on, I had consulted @peterjc about the project.
