+++ bamnostic/bgzf.py
@@ -10,7 +10,7 @@ license.  Please see the LICENSE file that should have been included
 as part of this package.
 
 Some methods are modified versions of their conterparts
-within the BioPython.bgzf module. Below is the Copyright and licensing 
+within the BioPython.bgzf module. Below is the Copyright and licensing
 for those parts.
 Copyright (c) 2010-2015 by Peter Cock.
 
@@ -33,10 +33,10 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 
 Description: Read and write BGZF compressed files (the GZIP variant used in BAM).
-Significant changes were made to the original BGZF module, produced by 
+Significant changes were made to the original BGZF module, produced by
 Peter Cock. Aside from adding an LRU dictionary, the new BGZF module can read
 BAM files directly, decompressing and unpacking the byte-encoded data structure
-outlined in the BAM_ format. 
+outlined in the BAM_ format.
 
 .. _BAM: https://samtools.github.io/hts-specs/SAMv1.pdf
 
@@ -63,45 +63,47 @@ _PY_VERSION = sys.version
 if _PY_VERSION.startswith('2'):
     from io import open
 
+
 def _format_warnings(message, category, filename, lineno, file=None, line=None):
     """ Warning formatter
-    
+
     Args:
         message: warning message
         category (str): level of warning
         filename (str): path for warning output
         lineno (int): Where the warning originates
-    
+
     Returns:
         Formatted warning for logging purposes
 
     """
     return ' {}:{}:{}: {}\n'.format(category.__name__, filename, lineno, message)
 
+
 warnings.formatwarning = _format_warnings
 
 # Constants used in BGZF format
-_bgzf_magic = b"\x1f\x8b\x08\x04" # First 4 bytes of BAM file
-_bgzf_header = b"\x1f\x8b\x08\x04\x00\x00\x00\x00\x00\xff\x06\x00\x42\x43\x02\x00" # Ideal GZIP header
-_bgzf_eof = b"\x1f\x8b\x08\x04\x00\x00\x00\x00\x00\xff\x06\x00BC\x02\x00\x1b\x00\x03\x00\x00\x00\x00\x00\x00\x00\x00\x00" # 28 null byte signature at the end of a non-truncated BAM file
-_bytes_BC = b"BC" # "Payload" or Subfield Identifiers 1 & 2 of GZIP header
+_bgzf_magic = b"\x1f\x8b\x08\x04"  # First 4 bytes of BAM file
+_bgzf_header = b"\x1f\x8b\x08\x04\x00\x00\x00\x00\x00\xff\x06\x00\x42\x43\x02\x00"  # Ideal GZIP header
+_bgzf_eof = b"\x1f\x8b\x08\x04\x00\x00\x00\x00\x00\xff\x06\x00BC\x02\x00\x1b\x00\x03\x00\x00\x00\x00\x00\x00\x00\x00\x00"  # 28 null byte signature at the end of a non-truncated BAM file
+_bytes_BC = b"BC"  # "Payload" or Subfield Identifiers 1 & 2 of GZIP header
 
 
 def _as_bytes(s):
     """ Used to ensure string is treated as bytes
-    
+
     The output
-    
+
     Args:
         s (str): string to convert to bytes
-    
+
     Returns:
         byte-encoded string
-    
+
     Example:
-        >>> str(_as_bytes('Hello, World').decode()) # Duck typing to check for byte-type object 
+        >>> str(_as_bytes('Hello, World').decode()) # Duck typing to check for byte-type object
         'Hello, World'
-    
+
     """
[…]
