+++ bamnostic/core.py
@@ -28,7 +28,7 @@ _PY_VERSION = sys.version
 
 Cigar = namedtuple('Cigar', ('op_code', 'n_op', 'op_id', 'op_name'))
 """``namedtuple`` for handling CIGAR data
-    
+
     Args:
         op_code (int): CIGAR operation index
         n_op (int): number of operations for a given op_code
@@ -39,8 +39,8 @@ Cigar = namedtuple('Cigar', ('op_code', 'n_op', 'op_id', 'op_name'))
 # The BAM format uses byte encoding to compress alignment data. One such
 # compression is how CIGAR operations are stored: they are stored and an
 # array of integers. These integers are mapped to their respective
-# operation identifier. Below is the mapping dictionary. 
-#_CIGAR_OPS = {
+# operation identifier. Below is the mapping dictionary.
+# _CIGAR_OPS = {
 #    'M' : ('BAM_CMATCH', 0),
 #    'I' : ('BAM_CINS', 1),
 #    'D' : ('BAM_CDEL', 2),
@@ -59,35 +59,35 @@ _SEQ_KEY = '=ACMGRSVTWYHKDBN'
 
 def offset_qual(qual_string):
     """ Offsets the ASCII-encoded quality string to represent PHRED score.
-    
+
     Every base that is in the alignment is assigned a Phred score. A Phred
     score (:math:`Q`) is defined as :math:`Q=-10\log_{10}P`, where :math:`P`
     is the base-calling error probability. Phred quality scores tend range
     from 10 to 60. These qualities are then offset by 33 and ASCII-encoded
     for readability and storage.
-    
+
     Args:
         qual_string (:py:obj:`str` or :py:obj:`bytes`): Phred quality scores without offset
-    
+
     Returns:
         (str): ASCII-encoded Phred scores offest by adding 33 to base score.
-    
+
     Examples:
         >>> qual_score = '\x1b\x1b\x1b\x16\x1b\x1b\x1b\x1a\x1b\x1b\x1b\x1b\x1b\x1b\x1b\x1b\x17\x1a\x1a\x1b\x16\x1a\x13\x1b\x1a\x1b\x1a\x1a\x1a\x1a\x1a\x18\x13\x1b\x1a'
         >>> ''.join(offset_qual(qual_score))
         '<<<7<<<;<<<<<<<<8;;<7;4<;<;;;;;94<;'
-    
+
     """
     def offset(base_qual):
         """Offsets the given byte code by 33 and returns the ASCII
         representation of it.
-        
+
         Args:
             base_qual (str): a single byte-encoded base score
-        
+
         Returns:
             ASII-encoded, offsetted representation of the base quality
-        
+
         Example:
             >>> offset('\x1b')
             '<'
@@ -96,8 +96,9 @@ def offset_qual(qual_string):
             return chr(base_qual + 33)
         except TypeError:
             return chr(ord(base_qual) + 33)
-    
+
     return map(offset, qual_string)
+
 
 # compiled/performant struct objects
 _unpack_refId_pos = struct.Struct('<2i').unpack
@@ -110,7 +111,7 @@ _unpack_array = struct.Struct('<si').unpack
 
 class AlignmentFile(bgzf.BgzfReader, bgzf.BgzfWriter):
     """Wrapper to allow drop in replacement for BAM functionality in a ``pysam``-like API.
-    
+
     Args:
         filepath_or_object (str | :py:obj:`file`): the path or file object of the BAM file
         mode (str): Mode for reading. BAM files are binary by nature (default: 'rb').
@@ -124,21 +125,21 @@ class AlignmentFile(bgzf.BgzfReader, bgzf.BgzfWriter):
         require_index (bool): require the presence of an index file or raise (default: False)
         duplicate_filehandle (bool): Not implemented. Raises warning if True.
[…]

[…]
