+++ setup.py
@@ -7,7 +7,7 @@ def readme():
 
 setup(
     name='bamnostic',
-    version='0.8.5',
+    version='0.8.6b1',
     description='Pure Python, OS-agnostic Binary Alignment Map (BAM) random access and parsing tool',
     long_description=readme(),
     url='https://github.com/betteridiot/bamnostic/',
