+++ setup.cfg
@@ -1,2 +1,2 @@
-[pycodestyle]
+[pep8]
 max-line-length = 140
