### 2018-06-20 865a87d
Correcting pep8 warnings and errors
