+++ setup.cfg
@@ -1,2 +1,2 @@
 [pep8]
-max-line-length = 140
+ignore = E501,E402
