+++ README.rst
@@ -74,11 +74,8 @@ Quickstart
 Bamnostic is meant to be a reduced drop-in replacement for
 `pysam <https://github.com/pysam-developers/pysam>`__. As such it has
 much the same API as ``pysam`` with regard to BAM-related operations.
-
-.. note:: the ``pileup()`` method is not supported at this time.
-
-Importing ``bamnostic``
-~~~~~~~~~~~~~~~~~~~~~~~
+**Note**: the ``pileup()`` method is not supported at this time. ###
+Importing
 
 .. code:: python
 
@@ -87,7 +84,7 @@ Importing ``bamnostic``
 Loading your BAM file
 ~~~~~~~~~~~~~~~~~~~~~
 
-Bamnostic comes with an *example BAM* (and respective BAI) file just to
+Bamnostic comes with an example BAM (and respective BAI) file just to
 play around with the output. Note, however, that the example BAM file
 does not contain many reference contigs. Therefore, random access is
 limited. This example file is made availble through
@@ -96,7 +93,7 @@ within the package.
 
 .. code:: python
 
-    >>> bam = bs.AlignmentFile(bs.example_path, 'rb')
+    bam = bs.AlignmentFile(bs.example_path, 'rb')
 
 Get the header
 ~~~~~~~~~~~~~~
@@ -108,8 +105,8 @@ length tuple values.
 
 .. code:: python
 
-    >>> bam.header
-    {0: ('chr1', 1575), 1: ('chr2', 1584)}
+    bam.header
+    >>> {0: ('chr1', 1575), 1: ('chr2', 1584)}
 
 Data validation through ``head()``
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -120,143 +117,53 @@ Data validation through ``head()``
     [EAS56_57:6:190:289:82  69  chr1    100 0   *   =   100 0   CTCAAGGTTGTTGCAAGGGGGTCTATGTGAACAAA <<<7<<<;<<<<<<<<8;;<7;4<;<;;;;;94<; MF:C:192,
      EAS56_57:6:190:289:82  137 chr1    100 73  35M =   100 0   AGGGGTGCAGAGCCGAGTCACGGGGTTGCCAGCAC <<<<<<;<<<<<<<<<<;<<;<<<<;8<6;9;;2; MF:C:64 Aq:C:0  NM:C:0  UQ:C:0  H0:C:1  H1:C:0]
 
-Getting the next read
-~~~~~~~~~~~~~~~~~~~~~
+Getting the first read
+~~~~~~~~~~~~~~~~~~~~~~
 
 .. code:: python
 
-    >>> print(next(bam))
-    EAS56_57:6:190:289:82	69	chr1	100	0	*	=	100	0	CTCAAGGTTGTTGCAAGGGGGTCTATGTGAACAAA	<<<7<<<;<<<<<<<<8;;<7;4<;<;;;;;94<;	MF:C:192
+    first_read = next(bam)
+    print(first_read)
+    >>> EAS56_57:6:190:289:82   69  chr1    100 0   *   =   100 0   CTCAAGGTTGTTGCAAGGGGGTCTATGTGAACAAA <<<7<<<;<<<<<<<<8;;<7;4<;<;;;;;94<; MF:C:192
 
-Exploring a read
-~~~~~~~~~~~~~~~~
-
-.. code:: python
-
-    # Using a more complex read
-    >>> for complex_read in bam:
-    ...     if complex_read.read_name == 'EAS218_4:1:48:9:409':
-    ...         print(complex_read)
-    ...         break
[…]
