### 2018-06-20 da7d641
fixing more pep8
