+++ setup.cfg
@@ -1,2 +1,5 @@
 [pep8]
 ignore = E501,E402
+
+[pycodestyle]
+ignore = E501,E402
