### 2018-06-20 4c4073a
Fixed code block duplication
