### 2018-06-20 374ca37
adding more to travis yml
