+++ .travis.yml
@@ -8,6 +8,10 @@ python:
   - "3.4"
   - "3.5"
   - "3.6"
-  - "3.7"
+  # PyPy versions
   - "pypy2.7"
   - "pypy3.5"
+# command to install dependencies
+install:
+  - pip install -r requirements.txt
+  - pip install .
