+++ .travis.yml
@@ -1,3 +1,7 @@
+os:
+  - linux
+  - windows
+
 language: python
 python:
   - "2.7"
@@ -8,10 +12,12 @@ python:
   - "3.4"
   - "3.5"
   - "3.6"
-  # PyPy versions
   - "pypy2.7"
   - "pypy3.5"
+
 # command to install dependencies
 install:
   - pip install -r requirements.txt
   - pip install .
+
+script: pytest
