+++ tests/test_build.py
@@ -6,7 +6,7 @@ def test_header():
     with bs.AlignmentFile(bs.example_bam, 'rb') as bam:
         expected = {0: ('chr1', 1575), 1: ('chr2', 1584)}
         observed = bam.header()
-        assert observed == expected, f'{type(expected)}\n{type(observed)}'
+        assert observed == expected
 
 
 def test_first_read():
