### 2018-06-20 1706772
reduced python versioning
