+++ .travis.yml
@@ -5,12 +5,6 @@ os:
 language: python
 python:
   - "2.7"
-  - "3.0"
-  - "3.1"
-  - "3.2"
-  - "3.3"
-  - "3.4"
-  - "3.5"
   - "3.6"
   - "pypy2.7"
   - "pypy3.5"
