### 2018-06-20 9afc067
Commented out pypy2.7 from travis.yml
the travis CI softlink to pypy2.7 seems to be failing.
