+++ .travis.yml
@@ -6,7 +6,7 @@ language: python
 python:
   - "2.7"
   - "3.6"
-  - "pypy2.7"
+  # - "pypy2.7"
   - "pypy3.5"
 
 # command to install dependencies
