### 2018-06-21 #779 @peterjc
Yes, @betteridiot got in touch with me about BAMnostic  be earlier this year - I've not been able to make time for any detailed input, but did suggest switching to a standard OSI approved open source license (since done).

My BGZF code referred to is in Biopython, see https://github.com/biopython/biopython/blob/master/Bio/bgzf.py

That BGZF code is now dual licensed (3-clause BSD or the legacy Biopython License Agreement), so having a derivate of that code here in BAMnostic under the 3-clause BSD license is absolutely fine from my point of view.

My pure-Python SAM/BAM code written on top of this still exists on branches, currently the latest code is here:
https://github.com/peterjc/biopython/tree/SamBam2015/Bio/Sequencing/SamBam

I have not had the time to polish this up to release ready status, and was wary of the continuously evolving nature of the SAM/BAM/CRAM file formats turning that into a continuous support burden. Currently the latest branch is here:

I would be technically able to review this, but the timing is awkward - I am away next week for a conference etc. Unless you mind waiting until mid July 2018, it would be better to ask someone else. Sorry.
