### 2018-06-21 48455fe
adding more Travis CI builds
