+++ .travis.yml
@@ -5,6 +5,8 @@ os:
 language: python
 python:
   - "2.7"
+  - "3.4"
+  - "3.5"
   - "3.6"
   # - "pypy2.7"
   - "pypy3.5"
