### 2018-06-21 aba1656
adding more Python versions to travis build
