### 2018-06-21 d44bf94
removing problematic python versions from travis build
