+++ .travis.yml
@@ -5,10 +5,10 @@ os:
 language: python
 python:
   - "2.7"
-  - "3.0"
-  - "3.1"
-  - "3.2"
-  - "3.3"
+#  - "3.0"
+#  - "3.1"
+#  - "3.2"
+#  - "3.3"
   - "3.4"
   - "3.5"
   - "3.6"
