+++ .travis.yml
@@ -13,7 +13,7 @@ python:
   - "3.5"
   - "3.6"
   - "3.7-dev"
-  # - "pypy2.7"
+#   - "pypy2.7"
   - "pypy3.5"
 
 # command to install dependencies
