### 2018-07-17 betteridiot/bamnostic#3 @peterjc
CSI indexing support as well as BAI indexing
Is there any plan to support CSI indexing? This adds a number of benefits over BAI which has a fixed binning tree and a hard coded maximum reference length, making it unsuitable for many genomes (e.g. lots of plants, some marsupials).
