### 2018-07-17 betteridiot/bamnostic#4 @peterjc
CRAM support?
The HTSlib team are pushing CRAM as the successor to BAM. This is a much more complicated format, so are there any plans to support that here?
