### 2018-07-17 betteridiot/bamnostic#5 @peterjc
BAM output support
The README says BAMnostic is *a pure Python, OS-agnositic Binary Alignment Map (BAM) file parser and random access tool.*

Are you intending to support BAM output, or is this only possible though a workaround like outputting SAM format and piping this though a command line tool like samtools?
