### 2018-07-17 betteridiot/bamnostic#5 @peterjc
This isn't that complicated, see e.g. my implementation at https://github.com/peterjc/biopython/blob/SamBam2015/Bio/Sequencing/SamBam/__init__.py#L1714
