### 2018-07-17 betteridiot/bamnostic#6 @peterjc
Support CIGARs with >65535 operations in BAM files
See https://github.com/samtools/hts-specs/issues/40 and the specification update in https://github.com/samtools/hts-specs/pull/227
