### 2018-07-18 #826 @pjotrp
@betteridiot we are starting review here. To expedite the review do you mind going through above list of check boxes and make sure they can be ticked (you can't tick them). Also check the PDF output carefully. Ping us here when you are done.
