### 2018-07-18 #826 @pjotrp
Also address @peterjc concerns raised https://github.com/openjournals/joss-reviews/issues/779#issuecomment-405429342 here:

One thing on the text: *"This is a significant limitation as no other Python implementation (besides pysam) can perform random access operation on BAM files."* maybe insert the rider "published" or "released" as there are several, including my old Python code can do this but has not been formally published or released, https://github.com/JohnLonginotto/pybam ("A simple, 100% python, BAM file reader.") and some work at https://github.com/nijibabulu/pypysam/

As a potential user I would also want to know about plans for CRAM support (logged as https://github.com/betteridiot/bamnostic/issues/4), CSI indexing (https://github.com/betteridiot/bamnostic/issues/3), and BAM output (https://github.com/betteridiot/bamnostic/issues/5).

I know this is a lot to ask, and perhaps out of scope for JoSS, but as a user I would worry about the long term support as the SAM/BAM file format continues to evolve, (e.g. the overdue fix for long CIGAR strings last year, important with long reads - logged as https://github.com/betteridiot/bamnostic/issues/6). In contrast, by building on top of the well supported HTSlib, pysam gets a lot of this maintenance work "for free", but a re-implementation would not (as I wrote earlier, this was a factor in me not formally releasing my own Python SAM/BAM code a years back).
