### 2018-07-19 #826 @betteridiot
The PR: betteridiot/bamnostic#7 addresses the first issue raised by @peterjc.

Also, a couple of these points were also addressed in my last comment in #779. I will reiterate it here though:

> Thank you greatly for your points. Regardless of @pjotrp decision to keep you on as the reviewer, you made great points that I would like to take the opportunity to address.
>
> Your points regarding BAM output and long CIGAR strings are definitely revisions that can (and should) be made to bamnostic. The CIGAR string revision is still a relatively recent change, and was not written into bamnostic yet. Thank you for pointing this out, and that pull request should come very soon. The BAM output point is a bigger issue, but still fair. For the purposes of the original project that birthed bamnostic, BAM output was not desired, and therefore fell victim to oversight. However, due to incorporating @peterjc BGZF module, the framework for connecting the pieces is there. They just have not been formally implemented. Making this change should be relatively straight-forward (much like you stated in the respective issue thread).
>
> I also agree with you that CRAM support would be a great boon for bamnostic. However, like you stated in betteridiot/bamnostic#4, it is a much more complex format that would require a non-trivial expenditure of time and resources to write up. Not saying that this feature is not in the future of bamnostic, but we believe that the present form of bamnostic could still be a very impactful tool for current pipelines and analysis. The same applies to CSI indexing. While it is a successor to BAI indexing, it has not permeated the domain to the degree that BAI has yet.
>
> With regard to the phrasing: "This is a significant limitation as no other Python implementation (besides pysam) can perform random access operation on BAM files.", your point is fair that published should be added.
>
> Additionally, in the vein of JoSS' purpose, this is an open-source project as well, and contributions are not only welcomed, they are encouraged. If the community would like to add features into bamnostic, pull requests are fully appreciated.
