### 2018-07-19 betteridiot/bamnostic#4 @betteridiot
I also agree with you that CRAM support would be a great boon for bamnostic. However, like you stated, it is a much more complex format that would require a non-trivial expenditure of time and resources to write up. Not saying that this feature is not in the future of bamnostic, but we believe that the present form of bamnostic could still be a very impactful tool for current pipelines and analysis.
