### 2018-07-19 betteridiot/bamnostic#3 @betteridiot
As stated in openjournals/joss-reviews#826, CSI indexing suffers the same fate as CRAM format: not that it is too complex, but it is still too new. While it is a successor to BAI indexing, it has not permeated the domain to the degree that BAI has yet. There are plans to incorporate it as the community grows around bamnostic, but we still believe bamnostic is impactful in its current state as well.
