### 2018-07-19 betteridiot/bamnostic#4 @peterjc
That's what I expected - I suggest adding something to the README file which should preempt m repeats of my question.
