### 2018-07-19 #826 @betteridiot
The points about CSI indexing and CRAM support were addressed similarly in betteridiot/bamnostic#3 and betteridiot/bamnostic#4 (respectively).
