### 2018-07-19 betteridiot/bamnostic#4 @betteridiot
Noted. Thank you
