### 2018-07-19 #826 @betteridiot
I will have to explicitly add a section on **Community guidelines** for all items to be checked off.
