### 2018-07-19 d2dc201
Added support for CIGAR strings > 65535
