### 2018-07-19 2b93eb3
Removed non-functional logic
