### 2018-07-19 betteridiot/bamnostic#6 @betteridiot
6e55d648ecfbf8a01332c3ef160c9be8b556c1e5 provides support for CIGARs >65535 operations
