### 2018-07-19 betteridiot/bamnostic#6 @betteridiot
This was tested on the example file provided by @lh3 in the samtools/hts-specs#40 thread
