### 2018-07-19 14747e7
Updated versioning documentation and added comments about lack of CRAM and CSI support at this time
