+++ README.md
@@ -68,7 +68,7 @@ Bamnostic is meant to be a reduced drop-in replacement for [pysam](https://githu
 >>> import bamnostic as bs
 ``` 
 
-### Loading your BAM file
+### Loading your BAM file (Note: CRAM and CSI formats are not supported at this time)
 Bamnostic comes with an example BAM (and respective BAI) file just to play around with the output. Note, however, that the example BAM file does not contain many reference contigs. Therefore, random access is limited. This example file is made availble through `bamnostic.example_bam`, which is a just a string path to the BAM file within the package.
 
 ```python
