### 2018-07-19 05b00e1
Added community guideline information
