+++ README.md
@@ -194,3 +194,8 @@ A simple seek call against the BAM file will put the head at the start of your r
 The common practice within the field of genomics/genetics when analyzing BAM files is to use the program known as [samtools](http://samtools.sourceforge.net/). The maintainers of samtools have done a tremendous job of providing distributions that work on a multitude of operating systems. While samtools is powerful, as a command line interface, it is also limited in that it doesn't really afford the ability to perform real-time dynamic processing of reads (without requiring many system calls to samtools). Due to its general nature and inherent readability, a package was written in Python called [pysam](https://github.com/pysam-developers/pysam). This package allowed users a very comfortable means to doing such dynamic processing. However, the foundation of these tools is built on a C-API called [htslib](https://github.com/samtools/htslib) and htslib cannot be compiled in a Windows environment. By extension, neither can pysam. 
 </br></br>
 In building a tool for genomic visualization, I wanted it to be platform agnostic. This is precisely when I found out that the tools I had planned to use as a backend did not work on Windows...the most prevalent operation system in the end-user world. So, I wrote **bamnostic**. As of this writing, bamnostic is OS-agnostic and written completely in Pure Python--requiring only the standard library (and `pytest` for the test suite). Special care was taken to ensure that it would run on all versions of CPython 2.7 or greater. Additionally, it runs in both stable versions of PyPy. While it may perform slower than its C counterparts, bamnostic opens up the science to a much greater end-user group. Lastly, it is lightweight enough to fit into any simple web server (e.g. [Flask](http://flask.pocoo.org/)), further expanding the science of genetics/genomics.
+
+
+***
+## Community Guidelines:
+Eagerly accepting PRs for improvements, optimizations, or features. For any questions or issues, please feel free to make a post to bamnostic's [Issue tracker](https://github.com/betteridiot/bamnostic/issues) on github.
