### 2018-07-19 #826 @betteridiot
Current version is 0.8.7b2
Community guidelines and versioning were added in https://github.com/betteridiot/bamnostic/commit/05b00e1297133bfaece68925e03ae59d1c5a16b9

Addressed @peterjc's comments on CRAM and CSI support is betteridiot/bamnostic#4 and betteridiot/bamnostic#3 (respectively) and added documentation about limitations in https://github.com/betteridiot/bamnostic/commit/14747e7135260c69a77b2908e187faaaacf95ef7

CIGARs > 65535 support was added in https://github.com/betteridiot/bamnostic/commit/6e55d648ecfbf8a01332c3ef160c9be8b556c1e5
