### 2018-07-19 #826 @betteridiot
At this point, I can confirm all items on the checklist can be checked off. The only point made by @peterjc  that has yet to be addressed with increased functionality has been on BAM output. This is feature that will be added in the very near future and this issue can be found at betteridiot/bamnostic#5
