### 2018-07-19 betteridiot/bamnostic#4 @betteridiot
14747e7135260c69a77b2908e187faaaacf95ef7 added documentation to the README.md Stating that CRAM & CSI support was not available at this time.
