### 2018-07-19 bc98248
updated README.rst
