### 2018-07-19 be6f40d
updated README.rst
