### 2018-07-20 faea893
Cleaned up bgzf.py header to include updated license information
As per https://github.com/biopython/biopython/commit/93ff47b5399b4e1d83815b2d075e2070ad98ba9f#diff-8f35ada1444d6f77e41c86b649ba88ec biopython.bgzf is goverened under the same 3-clause BSD as bamnostic.
