+++ bamnostic/bgzf.py
@@ -9,28 +9,10 @@ This code is part of the bamnostic distribution and governed by its
 license.  Please see the LICENSE file that should have been included
 as part of this package.
 
-Some methods are modified versions of their conterparts
-within the BioPython.bgzf module. Below is the Copyright and licensing
-for those parts.
+Some methods are modified versions of their counterparts within the BioPython.bgzf
+ module. BioPython BGZF is now under a 3-clause BSD license. The same license 
+found in this package. Refer to LICENSE for information.
 Copyright (c) 2010-2015 by Peter Cock.
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all
-copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-SOFTWARE.
 
 Description: Read and write BGZF compressed files (the GZIP variant used in BAM).
 Significant changes were made to the original BGZF module, produced by
