### 2018-07-20 betteridiot/bamnostic#5 @betteridiot
Make it more formal here that I am currently working on supporting BAM output with bamnostic. I will keep this issue open until that PR is complete
