### 2018-07-22 #826 @peterjc
BAM output https://github.com/betteridiot/bamnostic/issues/5 was a feature request, quite a logical extension and one which will increase the utility of BAMnostic, but I did not mean to suggest it should delay or block this JOSS publication.
