### 2018-07-22 #826 @betteridiot
Thank you @peterjc . I fully understood it as such. I was just being explicit with my intentions is all.
