+++ .travis.yml
@@ -5,7 +5,7 @@ os:
 language: python
 python:
   - "2.7"
-#  - "3.0"
+  - "3.0"
 #  - "3.1"
 #  - "3.2"
 #  - "3.3"
@@ -13,7 +13,7 @@ python:
   - "3.5"
   - "3.6"
   - "3.7-dev"
-#   - "pypy2.7"
+  - "pypy"
   - "pypy3.5"
 
 # command to install dependencies
