### 2018-07-26 7cb613c
Update issue templates
