### 2018-08-07 #826 @luizirber
Hello @betteridiot! Great work on the software =]

Some comments and two small things:
- [ ] Version
- [ ] References

# Checklist comments

## General checks

- Version:
  - need to update to 0.8.9 (latest version on master)
  - also: tag versions in git so they show up in the releases tab https://github.com/betteridiot/bamnostic/releases

## Documentation

- Automated tests: while there two tests, they don't cover large chunks of the codebase and API (especially the `utils` module).
  ```
  $ pytest --cov
  Name                    Stmts   Miss  Cover
  -------------------------------------------
  bamnostic/__init__.py       4      0   100%
  bamnostic/bai.py          159     63    60%
  bamnostic/bgzf.py         506    280    45%
  bamnostic/core.py         291    146    50%
  bamnostic/utils.py        246    178    28%
  tests/__init__.py           0      0   100%
  tests/test_build.py        11      0   100%
  -------------------------------------------
  TOTAL                    1217    667    45%
  ```
  Bringing in the doctests is already a nice improvement (but two tests need
  to be fixed):
  ```
  $ pytest --doctest-modules --cov
  Name                    Stmts   Miss  Cover
  -------------------------------------------
  bamnostic/__init__.py       4      0   100%
  bamnostic/bai.py          159     25    84%
  bamnostic/bgzf.py         506    169    67%
  bamnostic/core.py         291    114    61%
  bamnostic/utils.py        246     57    77%
  docs/source/conf.py        27      0   100%
  tests/__init__.py           0      0   100%
  tests/test_build.py        11      0   100%
  -------------------------------------------
  TOTAL                    1244    365    71%
  ```

- Community guidelines: There are issue templates for bugs and features, and a comment
  at the end of the README. A plus would be to collect this info in a [CONTRIBUTING file](https://blog.github.com/2012-09-17-contributing-guidelines/).

## Software paper

- References: there are two links to PySAM (repo and an issue) that don't have DOIs. But I'm not sure how to properly cite this... @pjotrp

# Fixes

- on https://github.com/betteridiot/bamnostic#loading-your-bam-file-note-cram-and-csi-formats-are-not-supported-at-this-time
  code should be
  `>>> bam = bs.AlignmentFile(bs.example_bam, 'rb')`
  instead of
  `>>> bam = bs.AlignmentFile(bs.example_path, 'rb')`
