### 2018-08-08 dde1507
JoSS review fixes
Fixed documentation error of README.md
Fixed the two failing doctests by correcting raised exception output
updated versioning
