### 2018-08-08 #826 @betteridiot
betteridiot/bamnostic@fbd5e89f2296d78732e addresses the issues raises by @luizirber regarding the versions releases, the documentation error of `bs.example_bam`, and fixes to the two failing doctests.

The CONTRIBUTING.md file will take more time than these quick corrections.

Regarding the comment about the automated tests: @luizirber , are you requesting more testing coverage is required before publication? Or, do the doctests fulfill the needs of the review at this point?
