### 2018-08-08 #826 @luizirber
Thanks for the quick answer!

It is not required, I think it's already pretty good. Having 100% coverage is desirable but not that useful if you're writing tests just to hit the target, but I was a bit uncomfortable with only two tests. After I found the doctests this looked better, and I would suggest (but not require :smile_cat:) that you add a few more tests, but up to you.

I checked the `version` box above, last question is about the references, but I'll wait @pjotrp for that.
