### 2018-08-08 #826 @pjotrp
@peterjc how would you like the paper to cite PySAM?
