### 2018-08-08 #826 @betteridiot
I know that according to [APA citations style guides](https://owl.purdue.edu/owl/research_and_citation/apa_style/apa_formatting_and_style_guide/reference_list_electronic_sources.html), when the DOI is not available for electronic resources, the reference should cite the URL (as done in [paper.bib](https://github.com/betteridiot/bamnostic/blob/master/docs/paper/paper.bib)).

Furthermore, the maintainers of PySAM requests that the original SAM/BAM format paper be cited in its stead (as done in [paper.bib](https://github.com/betteridiot/bamnostic/blob/master/docs/paper/paper.bib)). This is also seen in [pysam's Bioconda page](https://bioconda.github.io/recipes/pysam/README.html), where the DOI points back to the SAM/BAM format paper again.
