### 2018-08-08 9f424ec
added a contributing file
Community guidelines , code of conduct, and CONTRIBUTING.md file
