### 2018-08-08 d826c3e
Added in the Our Promise section
