### 2018-08-08 1a5ea58
Update CONTRIBUTING.md
