### 2018-08-08 9dda45b
Update CONTRIBUTING.md
