### 2018-08-08 6bbe4cd
Update CONTRIBUTING.md
