### 2018-08-08 #826 @betteridiot
https://github.com/betteridiot/bamnostic/pull/13 adds a CONTRIBUTING.md and Contributor Covenant Code of Conduct to the project. The initial commit had some typos in the CONTRIBUTING.md file. These were addressed in https://github.com/betteridiot/bamnostic/commit/6bbe4cd8d8caeee568b30766799549d3691d1965
