### 2018-08-08 3c074df
Update CODE_OF_CONDUCT.md
