+++ README.rst
@@ -329,7 +329,7 @@ github.
    :target: https://bamnostic.readthedocs.io/en/latest/?badge=latest
 .. |Maintainability| image:: https://api.codeclimate.com/v1/badges/d7e36e72f109c598c86d/maintainability
    :target: https://codeclimate.com/github/betteridiot/bamnostic/maintainability
-.. |DOI| image:: https://zenodo.org/badge/121782433.svg
-   :target: https://zenodo.org/badge/latestdoi/121782433
+.. |DOI| image:: https://zenodo.org/badge/DOI/10.5281/zenodo.1341915.svg
+   :target: https://doi.org/10.5281/zenodo.1341915
 .. |status| image:: http://joss.theoj.org/papers/9952b35bbb30ca6c01e6a27b80006bd8/status.svg
    :target: http://joss.theoj.org/papers/9952b35bbb30ca6c01e6a27b80006bd8
