### 2018-08-08 #826 @betteridiot
For expediency sake, the updated release (0.8.10c1) was updated to [Zenodo](https://zenodo.org/record/1341915) under the DOI: 10.5281/zenodo.1341915
