### 2018-08-08 0f7ad9b
Update README.rst
