### 2018-08-09 a368aa0
updated README.rst
