+++ README.rst
@@ -1,5 +1,15 @@
-|Build Status| |noarch| |Conda Version| |PyPI version| |License| |Conda
-Downloads| |Documentation Status| |Maintainability| |DOI| |status|
+`Build Status <https://travis-ci.org/betteridiot/bamnostic>`__
+`noarch <https://circleci.com/gh/conda-forge/bamnostic-feedstock>`__
+`Documentation
+Status <https://bamnostic.readthedocs.io/en/latest/?badge=latest>`__
+`Conda Version <https://anaconda.org/conda-forge/bamnostic>`__ `PyPI
+version <https://badge.fury.io/py/bamnostic>`__
+`Maintainability <https://codeclimate.com/github/betteridiot/bamnostic/maintainability>`__
+
+`status <http://joss.theoj.org/papers/9952b35bbb30ca6c01e6a27b80006bd8>`__
+`DOI <https://doi.org/10.5281/zenodo.1341959>`__
+`License <https://github.com/betteridiot/bamnostic/blob/master/LICENSE>`__
+`Conda Downloads <https://anaconda.org/conda-forge/bamnostic>`__
 
 BAMnostic
 =========
@@ -27,44 +37,44 @@ Through the ``conda`` package manager (`Anaconda Cloud <https://anaconda.org/con
 
 .. code:: bash
 
-    # first, add the conda-forge channel to your conda build
-    conda config --add channels conda-forge
+   # first, add the conda-forge channel to your conda build
+   conda config --add channels conda-forge
 
-    # now bamnostic is available for install
-    conda install bamnostic
+   # now bamnostic is available for install
+   conda install bamnostic
 
 Through the Python Package Index (`PyPI <https://pypi.org/>`__)
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 .. code:: bash
 
-    pip install bamnostic
+   pip install bamnostic
 
-    # or, if you don't have superuser access
-    pip install --user bamnostic
+   # or, if you don't have superuser access
+   pip install --user bamnostic
 
 Through pip+Github
 ~~~~~~~~~~~~~~~~~~
 
 .. code:: bash
 
-    # again, use --user if you don't have superuser access
-    pip install -e git+https://github.com/betteridiot/bamnostic.git
+   # again, use --user if you don't have superuser access
+   pip install -e git+https://github.com/betteridiot/bamnostic.git
 
-    # or, if you don't have superuser access
-    pip install --user -e git+https://github.com/betteridiot/bamnostic.git
+   # or, if you don't have superuser access
+   pip install --user -e git+https://github.com/betteridiot/bamnostic.git
 
 Traditional GitHub clone
 ~~~~~~~~~~~~~~~~~~~~~~~~
 
 .. code:: bash
 
-    git clone https://github.com/betteridiot/bamnostic.git
-    cd bamnostic
-    pip install -e .
+   git clone https://github.com/betteridiot/bamnostic.git
+   cd bamnostic
+   pip install -e .
 
-    # or, if you don't have superuser access
-    pip install --user -e .
+   # or, if you don't have superuser access
+   pip install --user -e .
 
 --------------
 
@@ -79,7 +89,7 @@ Importing
 
 .. code:: python
 
[…]
