### 2018-08-09 #826 @betteridiot
Thank you @pjotrp for keeping the review process moving. Thank you @peterjc  and @luizirber for your insights. Lastly, thank you JoSS for this outlet.
