### 2018-08-09 #826 @remills
Agreed, this process was great and we will definitely consider publishing here again in the future!
